clean!

+++ src/game/Landscape.java
@@ -257,7 +257,7 @@ public class Landscape extends GenLandTable
 		return _tile_type_procs[tile.GetTileType().ordinal()].get_tile_track_status_proc.applyAsInt(tile, mode);
 	}
 
-	static void ChangeTileOwner(TileIndex tile, byte old_player, byte new_player)
+	static void ChangeTileOwner(TileIndex tile, int old_player, int new_player)
 	{
 		_tile_type_procs[tile.GetTileType().ordinal()].change_tile_owner_proc.apply(tile, PlayerID.get( old_player ), PlayerID.get( new_player) );
 	}
